fixed some warnings

+++ Engine/Location.cpp
@@ -4,7 +4,7 @@ int Location::Distance(Location loc)
 {
 	int _x = this->x - loc.x;
 	int _y = this->y - loc.y;
-	return sqrt((_x*_x) + (_y*_y));
+	return int(sqrt((_x*_x) + (_y*_y)));
 }
 
 int Location::Distance(int _x, int _y)
